mac-ca: do keychains in parallel
Instead of going through each keychain individually, this gets the system
certificates from each keychain in parallel then returns the results as a
batch.  This uses more resources but finishes in ~2s instead of ~3s.

Note that this is called twice, once to set up Electron main process HTTPS
clients, and once for the VM.

Signed-off-by: Mark Yen <mark.yen@suse.com>