RPM spec: don't require pass
Since `pass` requires some setup, don't require it if `libsecret` is
available (which is more likely on a full desktop environment).

This is useful because the `docker` CLI, as well as things like `helm`,
assume that `pass` is set up if the command is found, because it is
typically only manually installed.

Unfortunately this cannot be implemented for Debian packages at this time
as that isn't implemented in `debbuild`; see
https://github.com/debbuild/debbuild/issues/1 for details.  To make this
work we'd have to have explicit DSC files instead.

Signed-off-by: Mark Yen <mark.yen@suse.com>

# Conflicts:
#	packaging/linux/rancher-desktop.spec