Scripts: with node-license-check with yarn plugin
Yarn auto inserts dependencies to `node-gyp` when it sees a `.gyp` file in
the package; however, we actually use `node-gyp-build` instead, so that
gets avoided (it's still used at build time, but not when running).  So we
need to customize the plugin to remove that link before checking the
licenses.

Note that we only do so for known packages, as we need to ensure that we do
not bring in a hard dependency on `node-gyp` from elsewhere.

Signed-off-by: Mark Yen <mark.yen@suse.com>