Diagnostics: connectivity: add settings for check interval
This adds a setting to adjust how often to do the network connectivity
check, in milliseconds, where setting it to zero disables the check.  Also
update the check itself to use lower-level APIs to avoid the issue with
Electron not handling redirects correctly (which seems to actually cause
the check to always pass).

Signed-off-by: Mark Yen <mark.yen@suse.com>