package.json: bump electron-builder to 26.2.0
This version is still @next instead of @latest, but it picks up the
relevant app-builder-lib changes to pull in the newer node-abi so we can
upgrade electron.  Note that this also removes the pin we previously had
on node-abi so we can use the newer version.  Just bumping the pin is not
sufficient as there are API-related changes elsewhere.

The bump to electron-updater is needed to keep the two in sync.

Signed-off-by: Mark Yen <mark.yen@suse.com>