Pull images before running them in bats/tests/containers/switch-engines.bats
This allows us to use --quiet to avoid flooding the logs with pull progress
messages.

Might be coincidence, but after making this change the test no longer failed
with "context deadline exceeded" for me on Windows. I suspect the slow pull
performance may have contributed to the likelihood of hitting the timeout.

Also adds --quiet option to some other places that pull without it.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>