Screenshots: Override containers/volumes before navigating
We set up the vuex store for containers and volumes when we navigate the
page; this means that we need to set up the override _before_ that happens,
otherwise we end up running the real code instead of the mocks.  This is
especially an issue because the real code will take longer, which means it
will revert any changes the mocks do.  This is particularly a problem on
slow machines, as that increases the chances we lose the race.

Signed-off-by: Mark Yen <mark.yen@suse.com>