Screenshots: override images list
To avoid runtime dependencies, mock out the images list (like we do for the
containers and volumes lists) and use that instead of real data for the
screenshots.  Note that we simply ignore the image processor; they still
run in the background.

Because the images list hasn't been converted to using the same RDX/vuex
based backend, the override is somewhat uglier here.
Signed-off-by: Mark Yen <mark.yen@suse.com>